Add API Auth to all Routes

+++ API/equipment/routes.py
@@ -1,5 +1,5 @@
 from flask import Blueprint, request, jsonify
-from API import db
+from API import db, bcrypt
 from API.models import Equipment
 from ..utils import shop_login_required
 from ..serializer import serialize_equipment
@@ -42,12 +42,28 @@ def fetch_all_equipments(current_user, public_id):
         :param public_id: Barbershop public_ID
         :return: 401, 200
     """
+    equipment_cost = 0
+    faulty_equipment = 0
     if current_user.public_id != public_id:
         return jsonify(dict(message="Not allowed")), 401
     all_equipments = []
-    for equipment in current_user.equipments:
+    equipments_query = current_user.equipments.order_by(Equipment.bought_on)
+    for equipment in equipments_query:
+        equipment_cost += equipment.price
+        if equipment.faulty:
+            faulty_equipment += 1
         all_equipments.append(serialize_equipment(equipment))
-    return jsonify(dict(equipments=all_equipments)), 200
+    oldest_equipment = equipments_query[0].equipment_name
+    newest_equipment = current_user.equipments.order_by(Equipment.bought_on.desc()).first()
+    return jsonify(dict(
+        equipments=all_equipments,
+        stats={
+            "cost": equipment_cost,
+            "faulty": faulty_equipment,
+            "oldest": oldest_equipment,
+            "newest": newest_equipment.equipment_name
+    }
+    )), 200
 
 
 @equipment_blueprint.route("/faulty/<int:equipment_id>", methods=["PUT"])
@@ -64,6 +80,10 @@ def mark_as_faulty(current_user, equipment_id):
         return jsonify(message="Not Found"), 404
     if current_user.public_id != equipment.shop.public_id:
         return jsonify(dict(message="Not allowed")), 401
+
+    data = request.get_json()
+    if not bcrypt.check_password_hash(current_user.password, data["password"].strip()):
+        return jsonify(dict(message="Incorrect Password")), 401
 
     equipment.faulty = True
     db.session.commit()
@@ -82,8 +102,13 @@ def remove_equipment(current_user, equipment_id):
     equipment = Equipment.query.filter_by(id=equipment_id).first()
     if not equipment:
         return jsonify(message="Not Found"), 404
+
     if current_user.public_id != equipment.shop.public_id:
         return jsonify(dict(message="Not allowed")), 401
+
+    data = request.get_json()
+    if not bcrypt.check_password_hash(current_user.password, data["password"].strip()):
+        return jsonify(dict(message="Incorrect Password")), 401
 
     db.session.delete(equipment)
     db.session.commit()
